Eduplan/rsp-skolni-system-qa-backend/cvut.fel.rsp/src/main/java/cvut/fel/rsp/eduaplan/model/Group.java
package cvut.fel.rsp.eduaplan.model;

import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
import javax.persistence.*;
import javax.validation.constraints.NotNull;
import java.util.Set;

@Entity
@Table(name = "group_")
@JsonIgnoreProperties(ignoreUnknown = true, value = {"hibernateLazyInitializer", "handler", "events"})
public class Group extends BaseEntity implements IEntity {

    @NotNull
    @Column(name="name")
    private String name;

    @OneToMany(mappedBy = "group")
    @JsonIgnoreProperties(ignoreUnknown = true, value = {"hibernateLazyInitializer", "handler", "group"})
    private Set<Student> students;

    @OneToMany(mappedBy = "group")
    private Set<Event> events;


    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public Set<Student> getStudents() {
        return this.students;
    }

    public void setStudents(Set<Student> students) {
        this.students = students;
    }

    public Set<Event> getEvents() {
        return events;
    }

    public void setEvents(Set<Event> events) {
        this.events = events;
    }
}
